Move onDrop into Home page
